modify modules
