+++ tests/analysis/neighborhood_analysis_test.py
@@ -9,8 +9,6 @@ import pytest
 import ark.settings as settings
 from ark.analysis import neighborhood_analysis
 import test_utils
-from pytest_mock import MockerFixture
-from alpineer.test_utils import _make_blank_file
 
 
 def test_create_neighborhood_matrix():
@@ -316,93 +314,3 @@ def test_compute_mixing_score():
         cell_neighbors_mat, 'fov1', target_cells=['cell4'], reference_cells=['cell2'],
         cell_count_thresh=0, mixing_type='homogeneous')
     assert math.isnan(cold_score)
-
-
-def test_shannon_diversity():
-
-    props1 = np.array([0, 0, 1])
-    assert neighborhood_analysis.shannon_diversity(props1) == 0
-
-    props2 = np.array([0, 0.25, 0, 0.15, 0.6])
-    assert neighborhood_analysis.shannon_diversity(props2) == \
-           -(0.25*np.log2(0.25)+0.15*np.log2(0.15)+0.6*np.log2(0.6))
-
-
-def test_compute_neighborhood_diversity():
-
-    neighbor_counts = pd.DataFrame({
-        settings.FOV_ID: ['fov1', 'fov1'],
-        settings.CELL_LABEL: list(range(1, 3)),
-        settings.CELL_TYPE: ['cell1', 'cell2'],
-        'cell1': [1, 0],
-        'cell2': [1, 2],
-    })
-
-    # counts matrix instead of freq should raise an error
-    with pytest.raises(ValueError, match="Input must be frequency values."):
-        neighborhood_analysis.compute_neighborhood_diversity(neighbor_counts, settings.CELL_TYPE)
-
-    neighbor_freqs = pd.DataFrame({
-        settings.FOV_ID: ['fov1', 'fov1', 'fov1', 'fov2'],
-        settings.CELL_LABEL: [1, 2, 3, 1],
-        settings.CELL_TYPE: ['cell1', 'cell2', 'cell2', 'cell1'],
-        'cell1': [0.4, 0, 0.5, 0.3],
-        'cell2': [0.3, 1, 0.5, 0],
-        'cell3': [0.3, 0, 0, 0.7]
-    })
-
-    diversity_data = neighborhood_analysis.compute_neighborhood_diversity(
-        neighbor_freqs, settings.CELL_TYPE)
-
-    # check for shannon diversity column
-    assert f'diversity_{settings.CELL_TYPE}' in diversity_data.columns
-
-    # check every cell is in the new dataframe
-    assert diversity_data.shape[0] == neighbor_freqs.shape[0]
-
-    # check for high vs low diversity
-    assert diversity_data[f'diversity_{settings.CELL_TYPE}'].max() == \
-           diversity_data[f'diversity_{settings.CELL_TYPE}'][0]
-    assert diversity_data[f'diversity_{settings.CELL_TYPE}'].min() == \
-           diversity_data[f'diversity_{settings.CELL_TYPE}'][1]
-
-
-def test_neighborhood_diversity_analysis():
-
-    with tempfile.TemporaryDirectory() as temp_dir:
-        radius = 50
-        cell_type_cols = ['cell_meta_cluster', 'cell_cluster']
-        for col in cell_type_cols:
-            cell1_freqs = [0.7, 0, 0.5, 0.2]
-            cell2_freqs = [0.3, 1, 0.5, 0.8]
-
-            if col == cell_type_cols[1]:
-                cell1_freqs = [1, 0.9, 0.6, 0.35]
-                cell2_freqs = [0, 0.1, 0.4, 0.55]
-
-            neighbor_freqs = pd.DataFrame({
-                settings.FOV_ID: ['fov1', 'fov1', 'fov1', 'fov2'],
-                settings.CELL_LABEL: [1, 2, 3, 1],
-                col: ['cell1', 'cell2', 'cell2', 'cell1'],
-                'cell1': cell1_freqs,
-                'cell2': cell2_freqs,
-            })
-            neighbor_freqs.to_csv(os.path.join(
-                temp_dir, f"neighborhood_freqs-{col}_radius{radius}.csv"), index=False)
-
-        # test success
-        all_data = neighborhood_analysis.neighborhood_diversity_analysis(
-           neighbors_mat_dir=temp_dir, pixel_radius=radius, cell_type_columns=cell_type_cols)
-
-        # check for multiple cell cluster columns
-        assert np.isin(cell_type_cols, all_data.columns).all()
-
-        diversity_columns = [f"diversity_{col}" for col in cell_type_cols]
-        assert np.isin(diversity_columns, all_data.columns).all()
-
-        # check every cell is in the new dataframe
-        assert all_data.shape[0] == all_data.shape[0]
-
-        # check score calculation by cell cluster
-        for index, row in all_data.iterrows():
-            assert row.diversity_cell_meta_cluster != row.diversity_cell_cluster
